requirements_win.txt
shapely~=1.8.1
gdal==3.4.3
fiona==1.8.21
pyproj==3.4.0
six==1.16.0
rtree==1.0.1
pygeos==0.13
geopandas==0.11.1